implemented the new generate dp banner
