+++ src/components/ui/Success.tsx
@@ -53,6 +53,15 @@ const Success = ({ title, content }: SuccessProps) => {
             WhatsApp Group
           </Link>
         </div>
+
+        <div className="flex gap-1 mb-2">
+          <Link
+            href="/generate-dp"
+            className="text-[var(--color-green-550)]  font-semibold"
+          >
+            Create Your Conference DP
+          </Link>
+        </div>
       </div>
     </div>
   );
